Add autherization to event delete routes

+++ frontend/src/components/Reviews.jsx
@@ -166,15 +166,25 @@ export const Reviews = () => {
   };
 
   const handleDeleteClick = async () => {
-    try {
-      await axios.delete(
-        `http://localhost:5000/api/review/deleteReview/${selectedReview._id}`
-      );
-      setReviews(reviews.filter((r) => r._id !== selectedReview._id));
-      setAlert("delete");
-      setSnackbarOpen(true);
-    } catch (err) {
-      console.log("Error deleting review:", err);
+    const user = JSON.parse(localStorage.getItem("user"));
+    if (user && user.token) {
+      try {
+        await axios.delete(
+          `http://localhost:5000/api/review/deleteReview/${selectedReview._id}`,
+          {
+            headers: {
+              Authorization: `Bearer ${user.token}`,
+            },
+          }
+        );
+        setReviews(reviews.filter((r) => r._id !== selectedReview._id));
+        setAlert("delete");
+        setSnackbarOpen(true);
+      } catch (err) {
+        console.log("Error deleting review:", err);
+      }
+    } else {
+      console.log("User not logged in or invalid access token");
     }
     handleMenuClose();
   };
